refactor: make linter happy
